pyproject.toml
[project]
name = "spine_engine"
dynamic = ["version"]
authors = [{name = "Spine Project consortium", email = "spine_info@vtt.fi"}]
license = {text = "LGPL-3.0-or-later"}
description = "A package to run Spine workflows."
keywords = ["energy system modelling", "workflow", "optimisation", "database"]
readme = {file = "README.md", content-type = "text/markdown"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8.1"
dependencies = [
    "networkx>2.5.1",
    "datapackage>=1.15.2, <1.16",
    "jupyter_client>=6.0",
    "spinedb_api>=0.31.2",
    "pyzmq >=21.0",
]

[project.urls]
Repository = "https://github.com/spine-tools/spine-engine"

[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2", "wheel", "build"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "spine_engine/version.py"
version_scheme = "release-branch-semver"

[tool.setuptools]
zip-safe = false

[tool.setuptools.package-data]
spine_engine = ["execution_managers/spine_repl.jl"]

[tool.setuptools.packages.find]
exclude = [
    "bin*",
    "docs*",
    "fig*",
    "tests*",
]

[tool.coverage.run]
source = ["spine_engine"]
branch = true
concurrency = ["multiprocessing", "thread"]
disable_warnings = ["no-data-collected"]

[tool.coverage.report]
ignore_errors = true

[tool.black]
line-length = 120
exclude = '\.git|version.py'

[tool.isort]
profile = "black"
lines_between_sections = 0
force_sort_within_sections = true
honor_case_in_force_sorted_sections = true
skip = ["version.py"]
line_length = 120
known_first_party = ["spinedb_api"]
